refactor : 여러곳에서 사용되는 타입들 하나의 types.ts에 모아놓음[#25]
