+++ src/components/main/types.ts
@@ -10,3 +10,12 @@ export type TabItem = {
     | '더보기 버튼';
   image: string;
 };
+
+// ** 채팅 방 리스트 개별 아이템에 사용되는 타입 & 스토리북에서도 사용 **
+export type ChatRoomItemProps = {
+  title: string;
+  total: number;
+  numOfPeople: number;
+  registerTime: number;
+  chattingUrl: string;
+};
